+++ changelog.md
@@ -1,4 +1,6 @@
 ## Change Log
 
+* 2019-11-15
+  * Added fetchParam(), fetchParamCache()
 * 2019-05-24
   * Initial Release